requirements.txt
selenium==4.18.1
BeautifulSoup4==4.12.3
requests==2.31.0
html5lib==1.1
python-dotenv==1.0.1

